[fix] ui restore basket order

+++ framework/core/src/python/kungfu/yijinjing/utils.py
@@ -8,6 +8,8 @@ from collections import deque
 def prune_layout_files(base_dir, layout, mode):
     search_path = os.path.join(base_dir, "*", "*", "*", layout, mode, "*")
     for file in filter(os.path.isfile, glob.glob(search_path)):
+        if "renderer-app" in file and layout == "log":
+            continue
         try:
             os.remove(file)
         except:
